refactored oc-plugin

+++ pkg/cmd/log_options.go
@@ -5,7 +5,7 @@ import "time"
 type LogOptions struct {
 	ID     string `json:"_id"`
 	Index  string `json:"_index"`
-	Score  int    `json:"_score"`
+	Score  float64    `json:"_score"`
 	Source struct {
 		Timestamp time.Time `json:"@timestamp"`
 		Docker    struct {
